% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addWorksheetWithSanitizedName}
\alias{addWorksheetWithSanitizedName}
\title{Add worksheet with sanitized name}
\usage{
addWorksheetWithSanitizedName(workbook, event_name, ...)
}
\arguments{
\item{workbook}{name of workbook.}

\item{event_name}{event name.}

\item{...}{params of addWorksheet.}
}
\value{
adds worksheet.
}
\description{
Add worksheet with sanitized name
}
