% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_data_on_firestore}
\alias{delete_data_on_firestore}
\title{Delete data on Firestore}
\usage{
delete_data_on_firestore(db_endpoint, auth_token)
}
\arguments{
\item{db_endpoint}{Character. Path to a collection on Firestore.}

\item{auth_token}{Character. Token to access Firestore.}
}
\value{
Response. It tells you whose and when the data was deleted.
}
\description{
Delete data on Firestore
}
