% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_all_name_combos}
\alias{get_all_name_combos}
\title{Get all name combos}
\usage{
get_all_name_combos(accessToken, project_name = PROJECT_NAME)
}
\arguments{
\item{accessToken}{Firebase access token.}

\item{project_name}{Firebase Project ID.}
}
\value{
A vector of name combos.
}
\description{
Get all name combos
}
