% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ci_data_for_a_day}
\alias{get_ci_data_for_a_day}
\title{Get and clean data for a day (all events).}
\usage{
get_ci_data_for_a_day(
  PROJECT_NAME,
  accessToken,
  Day,
  SME,
  Instructor,
  pilot_vec
)
}
\arguments{
\item{PROJECT_NAME}{Character. Firebase project name.}

\item{accessToken}{Character. Firebase access token.}

\item{Day}{Character. For example, "Day1", "Day2", etc.}

\item{SME}{Character. Name of Subject Matter Expert.}

\item{Instructor}{Character. Name of Instructor.}

\item{pilot_vec}{Character. Names of Pilot 1 and Pilot 2.}
}
\value{
A dataframe with event data.
}
\description{
Get and clean data for a day (all events).
}
