% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ci_data_for_a_day_v2}
\alias{get_ci_data_for_a_day_v2}
\title{Version 2 of get interview data for a day}
\usage{
get_ci_data_for_a_day_v2(
  accessToken,
  matching_name,
  Day,
  project_name = PROJECT_NAME
)
}
\arguments{
\item{accessToken}{Firebase token.}

\item{matching_name}{Combination of SME_Instructor_Pilot1_Pilot2}

\item{Day}{Day of simulation}

\item{project_name}{Firebase Project ID}
}
\value{
Dataframe with all events of a day for which data was collected.
}
\description{
Version 2 of get interview data for a day
}
