% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ci_event_data_for_a_day}
\alias{get_ci_event_data_for_a_day}
\title{Get and clean single event data for a day.}
\usage{
get_ci_event_data_for_a_day(
  PROJECT_NAME,
  accessToken,
  Day,
  event_name,
  SME,
  Instructor,
  Pilot1,
  Pilot2
)
}
\arguments{
\item{PROJECT_NAME}{Character. Firebase project name.}

\item{accessToken}{Character. Firebase access token.}

\item{Day}{Character. For example, "Day1", "Day2", etc.}

\item{event_name}{Character. Name of the event for which data is requested.}

\item{SME}{Character. Name of Subject Matter Expert.}

\item{Instructor}{Character. Name of Instructor.}

\item{Pilot1}{Character. Name of Pilot 1.}

\item{Pilot2}{Character. Name of Pilot 2.}
}
\value{
A dataframe with event data.
}
\description{
Get and clean single event data for a day.
}
