% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data_from_firestore}
\alias{get_data_from_firestore}
\title{Get data from Firestore}
\usage{
get_data_from_firestore(db_endpoint, auth_token)
}
\arguments{
\item{db_endpoint}{Character. Path to a document on Firestore.}

\item{auth_token}{Character. Token to access Firestore.}
}
\value{
Response. Data in JSON format.
}
\description{
Get data from Firestore
}
