% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{post_data_to_firestore}
\alias{post_data_to_firestore}
\title{Post user data to Firestore}
\usage{
post_data_to_firestore(db_endpoint, data, auth_token)
}
\arguments{
\item{db_endpoint}{Character. Path to a document on Firestore.}

\item{data}{JSON. Data in json format.}

\item{auth_token}{Character. Token to access Firestore.}
}
\value{
Response. It tells you the details of the posted data.
}
\description{
Post user data to Firestore
}
