% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{post_event_data_ci}
\alias{post_event_data_ci}
\title{Send event data to Firestore}
\usage{
post_event_data_ci(
  PROJECT_NAME,
  accessToken,
  Day,
  SME,
  Instructor,
  Pilot1,
  Pilot2,
  aircraft,
  pilot1_status,
  pilot1_title,
  pilot2_status,
  pilot2_title,
  event_name,
  constructs_inputs_p1,
  text_inputs_p1,
  constructs_inputs_p2,
  text_inputs_p2,
  notes,
  event_start,
  event_end
)
}
\arguments{
\item{PROJECT_NAME}{Character. Name of Firebase project.}

\item{accessToken}{Character. Access token.}

\item{Day}{Character. In the format Day<num> where num is 1, 2, etc.}

\item{SME}{Character. Name of Subject Matter Expert.}

\item{Instructor}{Character. Name of Instructor.}

\item{Pilot1}{Character. Name of pilot separated by space.}

\item{Pilot2}{Character. Name of pilot separated by space.}

\item{aircraft}{Aircraft Type.}

\item{pilot1_status}{Character. PF or PM.}

\item{pilot1_title}{Character. Captain or First Officer.}

\item{pilot2_status}{Character. PF or PM.}

\item{pilot2_title}{Character. Captain or First Officer.}

\item{event_name}{Character. Name of event. Examples: "event1", "event_of_interest_candidate", "event_of_interest_instructor".}

\item{constructs_inputs_p1}{Character. Responses to 9 constructs by SME for the event for Pilot1.}

\item{text_inputs_p1}{Character. Text responses of SMEs for Pilot1.}

\item{constructs_inputs_p2}{Character. Responses to 9 constructs by SME for the event for Pilot2.}

\item{text_inputs_p2}{Character. Text responses of SMEs for Pilot2.}

\item{notes}{Character. SME notes about the event.}

\item{event_start}{Character. Start of an event.}

\item{event_end}{Character. End of an event.}
}
\value{
Sends data to Firestore and returns an object of class response.
}
\description{
Send event data to Firestore
}
