% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sanitized_sheet_name}
\alias{sanitized_sheet_name}
\title{Sanitize event name}
\usage{
sanitized_sheet_name(event_name)
}
\arguments{
\item{event_name}{name of workbook.}
}
\value{
sanitized event name.
}
\description{
Sanitize event name
}
