% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sign.in}
\alias{sign.in}
\title{Sign in}
\usage{
\method{sign}{`in`}(email, password, api_key)
}
\arguments{
\item{email}{email of user.}

\item{password}{password of user.}

\item{api_key}{firebase api key.}
}
\value{
signed in user.
}
\description{
Sign in
}
